style: Renamed VSTSRestBuilds to Tfs
The documentation has been updated to reflect the changes to the file name.
