+++ app/services/Tfs.js
@@ -1,96 +1,203 @@
-var request = require('../requests');
+const request = require('../requests');
 
-module.exports = function () {
-    var self = this,
-        makeUrl = function (url, params) {
-            var baseUrl = 'https://' + self.configuration.pat + '@' + self.configuration.accountname + '.visualstudio.com/' + self.configuration.accountname + '/_apis/build/builds?api-version=4.1-preview';
+/** 
+ * The service which provides build information by using the VSTS REST API
+ *  v2.0.
+ * @public
+ * @constructor
+ * @see https://www.visualstudio.com/en-us/docs/integrate/api/build/overview
+ */
+function VSTSRestBuilds() {
+  let basicAuth = null;
+  let instance = null;
+  let project = null;
+  let params = null;
 
-            if (self.configuration.queryparams)
-            {
-                baseUrl += self.configuration.queryparams;
-                
-            }
-            if (params) {
-                baseUrl += params;
-            }
-            return baseUrl;
-        },
-        makeRequest = function (url, callback) {
-          request.makeRequest({
-            url: url,
-            headers: {Accept: 'application/json'}
-          }, callback);
-        },
-        parseDate = function (dateAsString) {
-            if (dateAsString == null) return null;
-            return new Date(dateAsString);
-        },
-        forEachResult = function (body, callback) {
-            for (var i = 0; i < body.value.length; i++) {
-                callback(body.value[i]);
-            }
-        },
-        isNullOrWhiteSpace = function (string) {
-            if(!string) {
-                return true;
-            }
+  /**
+   * This object is the representation of resultFilter mentioned in the docs
+   * @private
+   * @see https://www.visualstudio.com/en-us/docs/integrate/api/build/builds
+   */
+  const resultFilter = Object.freeze({
+    succeeded: 'succeeded',
+    partiallySucceeded: 'partiallySucceeded',
+    failed: 'failed',
+    canceled: 'canceled',
+  });
 
-            return string === null || string.match(/^ *$/) !== null;
-        },
-        getStatus = function (statusText) {
-            if (statusText === "succeeded") return "Green";
-            if (statusText === "failed") return "Red";
-            if (statusText === "inProgress") return "Blue";
-            if (statusText === "stopped") return "Gray";
-            if (statusText === "partiallySucceeded") return "'#FFA500'";
+  /**
+   * This object defines the color scheme used.
+   * @private
+   */
+  const colorScheme = Object.freeze({
+    succeeded: 'Green',
+    partiallySucceeded: 'Yellow',
+    failed: 'Red',
+    canceled: 'Black',
+  });
 
-            return null;
-        },
-        simplifyBuild = function (res) {
-            return {
-                id: res.project.name + '|' + res.definition.name + '|' + res.buildNumber,
-                project: res.project.name,
-                definition: res.definition.name,
-                number: res.buildNumber,
-                isRunning: res.result=='inProgress',
-                startedAt: parseDate(res.startTime),
-                finishedAt: parseDate(res.finishTime),
-                requestedFor: res.requestedFor.displayName,
-                statusText: (typeof res.result === 'undefined' || res.result==null)?res.status:res.result,
-                status: getStatus((typeof res.result === 'undefined' || res.result==null)?res.status:res.result),
-                reason: res.reason,
-                url: res._links.web.href,
-                hasErrors: !isNullOrWhiteSpace(res.errors),
-                hasWarnings: !isNullOrWhiteSpace(res.warnings)
-            };
-        },
-        queryBuilds = function (callback) {
-            makeRequest(makeUrl(null, null), function (error, body) {
-                if (error) {
-                  callback(error);
-                  return;
-                }
+  /** This object is the representation of statusFilter mentioned in the docs
+   * @private
+   * @see https://www.visualstudio.com/en-us/docs/integrate/api/build/builds
+   */
+  const statusFilter = Object.freeze({
+    inProgress: 'inProgress',
+    completed: 'completed',
+    cancelling: 'cancelling',
+    postponed: 'postponed',
+    notStarted: 'notStarted',
+    all: 'all',
+  });
 
-                if (body.error) {
-                  callback(new Error(body.error.message.value));
-                  return;
-                }
+  /**
+   * @typedef {Object} Build
+   * @property {Date} startedAt Build start time
+   * @property {Date} finishedAt Build finish time
+   * @property {boolean} hasErrors Does the resulting build have errors?
+   * @property {boolean} hasWarnings Did the build give some warnings?
+   * @property {boolean} isRunning Is the build currently running?
+   * @property {string} id Unique ID of the build
+   * @property {string} number Build number
+   * @property {string} project Name of the project
+   * @property {string} reason Reason for building the project
+   * @property {string} requestedFor Name of the Requester
+   * @property {string} status The color to be used for displaying 
+   * @property {string} statusText The status of the build
+   * @property {string} url URL of the project
+   */
 
-                var builds = [];
+  /**
+   * It is a node-style callback.
+   * @callback buildsInfoRequestCallback
+   * @param {Error|null} err It is an instance of Error
+   * @param {Array<Build>} listOfBuilds It is an array of {@link Build}
+   */
 
-                forEachResult(body, function (res) {
-                    builds.push(simplifyBuild(res));
-                });
+  /**
+   * It exposes the API needed by the application to check the status of builds.
+   * @name check
+   * @function
+   * @public
+   * @instance
+   * @memberOf VSTSRestBuilds
+   * @param {buildsInfoRequestCallback} cb Callback which handles the
+   *  requested build information
+   */
+  this.check = (callback) => {
+    if (basicAuth && instance && project) {
+      getListOfBuilds(callback);
+      return;
+    }
+    callback('incomplete configuration');
+    return;
+  };
 
-                callback(error, builds);
-            });
-        };
+  /**
+   * @typedef {Object} VSTSRestBuildsConfiguration
+   * @property {string} accountname VS Team Services account
+   *  ({account}.visualstudio.com)
+   * @property {string} collection Team project ID or name
+   * @property {string} queryparams Additional queryparams to filter the data
+   *  and provide additional options
+   * @property {string} username Username
+   * @property {string} pat Personal Access Token with access to Builds
+   *  information
+   */
 
-    self.configure = function (config) {
-        self.configuration = config;
+  /**
+   * It exposes the API needed by the application to provide
+   *  configuration parameters.
+   * @name check
+   * @function
+   * @public
+   * @instance
+   * @memberOf VSTSRestBuilds
+   * @param {VSTSRestBuildsConfiguration} config Configuration parameters
+   */
+  this.configure = (config) => {
+    /**
+     * It exposes the configuration passed to the 
+     * configure instance method.
+     * @public
+     * @instance
+     */
+    this.configuration = config;
+    basicAuth = new Buffer(`${config.username}:${config.pat}`)
+      .toString('base64');
+    instance = config.accountname;
+    params = config.queryparams;
+    project = config.collection;
+
+    console.log(config);
+  };
+
+  /**
+   * @private
+   * @param {buildsInfoRequestCallback} cb Callback which handles the
+   *  requested build information
+   */
+  const getListOfBuilds = (callback) => {
+    const url = `https://${instance}.visualstudio.com/DefaultCollection/${project}/_apis/build/builds?api-version=2.0${params}`;
+    const options = {
+      url,
+      headers: {
+        Authorization: `Basic ${basicAuth}`,
+      },
+    };
+    request.makeRequest(options, (err, body) => {
+      transformData(err, body, callback);
+    });
+
+    /**
+     * Transforms the data received from the request to VSTS REST API
+     * @private
+     * @param {any} err If the value is truthy, it indicates an error has
+     *  occurred.
+     * @param {object} body It contains the response body from VSTS REST API
+     * @param {buildsInfoRequestCallback} cb Callback which handles the
+     *  requested build information
+     * @return {null}
+     */
+    const transformData = (err, body, callback) => {
+      if (err) {
+        callback(err);
+        return;
+      }
+      if (!(body && body.value)) {
+        callback('No values found');
+        return;
+      }
+      const transformedData = body.value.map(transformer);
+      callback(null, transformedData);
     };
 
-    self.check = function (callback) {
-        queryBuilds(callback);
+    /**
+     * The function transforms the data from VSTS API to
+     *  the accepted by callback 
+     * @name transformer
+     * @private
+     * @param {object} build individual build information object returned
+     *  from API
+     * @returns {Build} the object is in the format accepted by the application
+     */
+    const transformer = (build) => {
+      return {
+        finishedAt: new Date(build.finishTime),
+        hasErrors: build.result === resultFilter.failed,
+        hasWarnings: build.result === resultFilter.partiallySucceeded,
+        id: build.id,
+        isRunning: build.status === statusFilter.inProgress,
+        number: build.buildNumber,
+        project: build.definition.name,
+        reason: build.reason,
+        requestedFor: build.requestedFor.displayName,
+        startedAt: new Date(build.startTime),
+        status: colorScheme[resultFilter[build.result]],
+        statusText: build.status,
+        url: build.url,
+      };
     };
-};
+  };
+}
+
+module.exports = VSTSRestBuilds;
